init project with dev, stg, and prod environment
