+++ QueProject/QueProject/AppDelegate.swift
@@ -1,0 +1,16 @@
+//
+//  AppDelegate.swift
+//  QueProject
+//
+//  Created by TrongQue on 12/2/25.
+//
+
+import UIKit
+
+@main
+class AppDelegate: UIResponder, UIApplicationDelegate {
+    
+    func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
+        return true
+    }
+}
