+++ QueProject/QueProject/Assets.xcassets/AccentColor.colorset/Contents.json
@@ -1,0 +1,12 @@
+{
+  "colors" : [
+    {
+      "filename" : "AppIconDebug.png",
+      "idiom" : "universal"
+    }
+  ],
+  "info" : {
+    "author" : "xcode",
+    "version" : 1
+  }
+}
