+++ QueProject/QueProject/Configs/Configurations.swift
@@ -1,0 +1,8 @@
+//
+//  Configurations.swift
+//  QueProject
+//
+//  Created by TrongQue on 13/2/25.
+//
+
+import Foundation
