+++ QueProject/QueProject/Configs/Debug.xcconfig
@@ -1,0 +1,14 @@
+//
+//  Debug.xcconfig
+//  QueProject
+//
+//  Created by TrongQue on 12/2/25.
+//
+
+// Configuration settings file format documentation can be found at:
+// https://help.apple.com/xcode/#/dev745c5c974
+
+QUE_APP_NAME = QueProject (dev)
+QUE_APP_VERSION = 1.0
+QUE_APP_BUNDLE_ID = com.que.QueProject.Debug
+QUE_APP_ICON = AppIconDebug
