+++ QueProject/QueProject/Configs/Release.xcconfig
@@ -1,0 +1,14 @@
+//
+//  Release.xcconfig
+//  QueProject
+//
+//  Created by TrongQue on 12/2/25.
+//
+
+// Configuration settings file format documentation can be found at:
+// https://help.apple.com/xcode/#/dev745c5c974
+
+QUE_APP_NAME = QueProject
+QUE_APP_VERSION = 1.0
+QUE_APP_BUNDLE_ID = com.que.QueProject.Release
+QUE_APP_ICON = AppIconRelease
