+++ QueProject/QueProject/MyViewController.swift
@@ -1,0 +1,15 @@
+//
+//  MyViewController.swift
+//  QueProject
+//
+//  Created by TrongQue on 12/2/25.
+//
+
+import UIKit
+
+class MyViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+    }
+}
