--- github-api/ViewController.swift
@@ -1,19 +1,0 @@
-//
-//  ViewController.swift
-//  github-api
-//
-//  Created by LamLe on 2/13/25.
-//
-
-import UIKit
-
-class ViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view.
-    }
-
-
-}
-
